fady amir c4s14:Receive notifications about inactive ideas
I added date object in comment object to know 
the date of each comment.
I added two methods in idea :the first one "getAllIdeas" that gets all organizations and all the users in each organization and all ideas of each user and calls the second method "checkDate" that takes the idea and checks if it is active or not.
i have also added in the bootstrap class thread called "timer" that calls every 24-hours the class "callTimer" which calls the method "getAllIdeas" to check for all inactive ideas by calling checkIdea.